+++ protected/humhub/modules/user/models/Group.php
@@ -441,6 +441,11 @@ class Group extends ActiveRecord
         $newGroupUser->created_by = Yii::$app->user->id;
         $newGroupUser->is_group_manager = $isManager;
         if ($newGroupUser->save()) {
+            if ($user instanceof User) {
+                // Drop stale cached group relations/flags on the passed instance
+                // so a following isSystemAdmin()/getGroups() sees this membership.
+                $user->invalidateGroupCache();
+            }
             if ($this->notify_users && !Yii::$app->user->isGuest) {
                 if (!($user instanceof User)) {
                     $user = User::findOne(['id' => $user]);
@@ -471,6 +476,9 @@ class Group extends ActiveRecord
         }
 
         if ($groupUser->delete()) {
+            if ($user instanceof User) {
+                $user->invalidateGroupCache();
+            }
             if ($this->notify_users) {
                 if (!($user instanceof User)) {
                     $user = User::findOne(['id' => $user]);